[ndlorenc] Portfolio_view follows repo design pattern and fixed decimal place issues with viewing portfolio
